% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function generate data visualization. For a given state, the function show points that represents
accidents in the state during the year given as a input.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{it is the code of the state, class numeric.}

\item{year}{As before it can be one number (one year), class numeric.}
}
\value{
The ouput of this function is plot a map with points that represents accidents. However, if you write
wrong the state code then the function stop and show the message "invalid STATE number: ___".
}
\description{
This function generate data visualization. For a given state, the function show points that represents
accidents in the state during the year given as a input.
}
\examples{
\dontrun{fars_map_state(state.num=6, year=2013)}
\dontrun{fars_map_state(state.num=29, year=2014)}

}
