% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Is a string what represents the file's name, class character.}
}
\value{
This function return a data frame of the data that was imported. If you write
wrong the file's name (or the file does not exist) then function return the message "file __ does not exist".
}
\description{
This function reads data. The document to read has data from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (a csv file), which is a nationwide census providing the American
public yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.
The input of the function is the filename to import and return a data frame .
}
\examples{
\dontrun{fars_read(filename="accident_2014.csv.bz2")}

}
