% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function use the function fars_read one or more times according to the amount of inputs.
Moreover, the function create a column with mutate called year and select variables MONTH and year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{it can be one number (one year) or a vector of numbers (years), class numeric.}
}
\value{
The output of this function is a single data frame with columns MONTH and year or a list
with the same columns (in case that input it is a list), class data frame. However, if a year is invalid then function
return a warning message "invalid year: _".
}
\description{
This function use the function fars_read one or more times according to the amount of inputs.
Moreover, the function create a column with mutate called year and select variables MONTH and year.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
\dontrun{fars_read_years(2015)}

}
