% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function summarize years, binding separate dataset for each year in one data frame with bind_rows.
After, with the spread function change rows into column, where years are in columns and months are in rows.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{As before it can be one number (one year) or a vector of numbers (years), class numeric.}
}
\value{
The ouput of this function is a data frame where columns are years and rows are months, it is
important to considerar that the information of the data frame represent the number of accidents for each month and year.
}
\description{
This function summarize years, binding separate dataset for each year in one data frame with bind_rows.
After, with the spread function change rows into column, where years are in columns and months are in rows.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
\dontrun{fars_read_years(2015)}

}
