% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{my_min}
\alias{my_min}
\title{min function}
\usage{
my_min(data)
}
\arguments{
\item{data}{data是一个数据框格式的变量 \code{data}}

\item{...}{\code{...}}
}
\value{
返回一个数值型变量，其值为数据框所有值的最小值
}
\description{
This is a min function for dataframe.
}
\details{
输入一个数据框, 输出数据框的最小值
}
\seealso{
\code{\link[tools]{my_max}}
}
\author{
lantary
}
