% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nico_move_animals_in_comparmets2.R
\name{nico_move_animals_in_comparmets}
\alias{nico_move_animals_in_comparmets}
\title{Move animals among compartments}
\usage{
nico_move_animals_in_comparmets(movements, linha)
}
\arguments{
\item{movements}{dataframe with the movements}

\item{linha}{day of the movemnts}
}
\value{

}
\description{
this function move the cattle pig and small ruminand ccording with the day and  SEIR compartment
}
\examples{
#'
> nico_move_animals_in_comparmets(movements = movements, 1)
date event_type        from          to   host number From To pop_S_ori pop_E_ori pop_I_ori pop_R_ori pop_S_move pop_E_move pop_I_move pop_R_move
136    1          1 43015003571 43001002076 Bovine      3  236  1       100       100       100       100          1          0          2          0
}
